+++ src/grogupy/cli/tests/input0_k.py
@@ -118,7 +118,7 @@ magnopy_precision = None
 # add the simulation parameters to the magnopy file as comments
 magnopy_comments = True
 
-# save the Uppsala Atomistic Spin Dynamics software input files
+# save the UppASD Atomistic Spin Dynamics software input files
 # uses the outfolder and out_spin_moment
 save_UppASD = True
 # add the simulation parameters to the inpsd.dat file as