+++ CHANGELOG.md
@@ -27,4 +27,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Failed tqdm loading prints error message only once through MPI.
 - Using spin moment instead of magnetic moment to follow magnopy convention.
-- .fdf input is not enforced in the command line tool.
+- *.fdf* input is not enforced in the command line tool.